Refactor the 'sync-extensions' command out of lib.rs.
This commit refactors the `sync-extensions` command out of lib.rs into a
command implementation.
