+++ azure-functions/src/commands.rs
@@ -1,5 +1,7 @@
 mod init;
 mod run;
+mod sync_extensions;
 
 pub use self::init::*;
 pub use self::run::*;
+pub use self::sync_extensions::*;
